Minor improvements in eventlog, temperature, win-update

+++ agent/checks/temperature.ps1
@@ -1,17 +1,23 @@
 Function prestart()
 {
-	$env:CheckMKstateDir = $STATEDIR
-	& ($COMPDIR + 'coretemp\Core Temp.exe')
+	if ($isVM -eq $false) {
+		$env:CheckMKstateDir = $STATEDIR
+		& ($COMPDIR + 'coretemp\Core Temp.exe')
+	}
 }
 
 Function run()
 {
-	Send-Line "<<<temperature>>>"
-	$statfile = $STATEDIR + "coretemp.log"
-	Send-Line (Get-Content $statfile)
+	if ($isVM -eq $false) {
+		Send-Line "<<<temperature>>>"
+		$statfile = $STATEDIR + "coretemp.log"
+		Send-Line (Get-Content $statfile)
+	}
 }
 
 Function terminate()
 {
-	taskkill /IM "Core Temp.exe" /F
+	if ($isVM -eq $false) {
+		taskkill /IM "Core Temp.exe" /F
+	}
 }
